+++ src/common/reliability/background_segm/tests/rgb_test.jpg
